simpan

+++ database/migrations/2024_01_01_000004_create_mata_kuliah_table.php
@@ -1,29 +1,22 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    public function up(): void
-    {
+return new class extends Migration {
+    public function up(): void {
         Schema::create('mata_kuliah', function (Blueprint $table) {
             $table->id();
             $table->foreignId('kampus_id')->constrained('kampus')->cascadeOnDelete();
             $table->foreignId('kelas_id')->constrained('kelas')->cascadeOnDelete();
             $table->string('kode', 20)->unique();
             $table->string('nama', 150);
-            $table->integer('sks')->default(2);
+            $table->tinyInteger('sks')->default(2);
             $table->enum('jenis', ['teori', 'praktikum', 'teori_praktikum'])->default('teori');
             $table->string('dosen', 100)->nullable();
-            $table->integer('total_pertemuan')->default(16);
+            $table->tinyInteger('total_pertemuan')->default(16);
             $table->timestamps();
         });
     }
-
-    public function down(): void
-    {
-        Schema::dropIfExists('mata_kuliah');
-    }
+    public function down(): void { Schema::dropIfExists('mata_kuliah'); }
 };
